move resize_preview from dev/utils to addons/utils
