+++ src/olympia/devhub/tasks.py
@@ -542,39 +542,6 @@ def pngcrush_existing_preview(preview_id):
     # automatically if the task was called with set_modified_on_obj=[preview].
 
 
-@task
-@set_modified_on
-def resize_preview(src, preview_pk, **kw):
-    """Resizes preview images and stores the sizes on the preview."""
-    preview = Preview.objects.get(pk=preview_pk)
-    preview.sizes = {'thumbnail_format': amo.ADDON_PREVIEW_SIZES['thumbnail_format']}
-    thumb_dst, full_dst, orig_dst = (
-        preview.thumbnail_path,
-        preview.image_path,
-        preview.original_path,
-    )
-    log.info('[1@None] Resizing preview and storing size: %s' % thumb_dst)
-    try:
-        (preview.sizes['thumbnail'], preview.sizes['original']) = resize_image(
-            src,
-            thumb_dst,
-            amo.ADDON_PREVIEW_SIZES['thumbnail'],
-            format=amo.ADDON_PREVIEW_SIZES['thumbnail_format'],
-        )
-        (preview.sizes['image'], _) = resize_image(
-            src,
-            full_dst,
-            amo.ADDON_PREVIEW_SIZES['full'],
-        )
-        if not os.path.exists(os.path.dirname(orig_dst)):
-            os.makedirs(os.path.dirname(orig_dst))
-        os.rename(src, orig_dst)
-        preview.save()
-        return True
-    except Exception as e:
-        log.error('Error saving preview: %s' % e)
-
-
 def _recreate_images_for_preview(preview):
     log.info('Resizing preview: %s' % preview.id)
     try:
